refactored some things
