+++ src/modules/user/user.module.ts
@@ -1,6 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserController } from './user.controller';
-import { UserController } from './controller/user/user.controller';
+import { UserController } from './controller';
 import { UserService } from './service/user.service';
 
 @Module({
